+++ src/app.js
@@ -1,0 +1,5 @@
+// CSS
+require('./scss/core.scss')
+
+// JS
+require('./js/core.js')