+++ src/js/app.js
@@ -1,8 +1,0 @@
-// CSS
-require('../scss/core.scss')
-
-// JS
-import 'bootstrap';
-import 'tablesorter';
-window.$ = require('jquery');
-require('./core.js')